requirements.txt
flask
requests
colorama
python-dotenv